Switch webpack config from custom to CRA
