--- babel.config.js
@@ -1,4 +1,0 @@
-module.exports = {
-    presets: ["@babel/preset-env", "@babel/preset-react"],
-    plugins: [["@babel/plugin-proposal-class-properties"]],
-};
