+++ frontend/src/app.js
@@ -1,8 +1,5 @@
 import React from "react";
-import ReactDOM from "react-dom";
 import "@patternfly/react-core/dist/styles/base.css";
-import "./style.css";
-
 import { BrowserRouter as Router } from "react-router-dom";
 import { AppLayout } from "./components/app_layout";
 import { AppRoutes } from "./routes";
@@ -16,6 +13,4 @@ const App = () => {
         </Router>
     );
 };
-
-const rootElement = document.getElementById("root");
-ReactDOM.render(<App />, rootElement);
+export default App;
