+++ frontend/src/components/chroot_status.js
