+++ frontend/src/components/copr_builds_table.js
@@ -70,7 +70,7 @@ const CoprBuildsTable = () => {
 
     // Fetch data from dashboard backend (or if we want, directly from the API)
     function fetchData() {
-        fetch(`${apiURL}/copr-builds?page=${page}&per_page=20`)
+        fetch(`${process.env.REACT_APP_API_URL}/copr-builds?page=${page}&per_page=20`)
             .then((response) => response.json())
             .then((data) => {
                 jsonToRow(data);
