+++ frontend/src/components/forge_icon.js
