+++ frontend/src/components/project_info.js
@@ -37,7 +37,7 @@ const ProjectInfo = (props) => {
     };
 
     function checkValidProject() {
-        fetch(`${apiURL}/projects/${forge}/${namespace}/${repoName}`)
+        fetch(`${process.env.REACT_APP_API_URL}/projects/${forge}/${namespace}/${repoName}`)
             .then((response) => response.json())
             .then((data) => {
                 if (!data.error) {
