+++ frontend/src/components/releases.js
@@ -18,7 +18,7 @@ const ReleasesList = (props) => {
 
     // Fetch data from dashboard backend (or if we want, directly from the API)
     function fetchData() {
-        fetch(`/api/projects/${forge}/${namespace}/${repoName}/releases`)
+        fetch(`${process.env.REACT_APP_API_URL}/projects/${forge}/${namespace}/${repoName}/releases`)
             .then((response) => response.json())
             .then((data) => {
                 setReleaseList(data);
