+++ frontend/src/components/search_project.js
