+++ frontend/src/components/trigger_info.js
