--- package.json
@@ -1,44 +1,0 @@
-{
-    "name": "dashboard",
-    "version": "0.0.1",
-    "dependencies": {
-        "@patternfly/react-core": "^4.15.0",
-        "@patternfly/react-icons": "^4.3.0",
-        "@patternfly/react-styles": "^4.3.0",
-        "@patternfly/react-table": "^4.5.7",
-        "react": "^16.13.1",
-        "react-dom": "^16.13.1",
-        "react-router-dom": "^5.2.0"
-    },
-    "devDependencies": {
-        "@babel/core": "^7.10.2",
-        "@babel/plugin-proposal-class-properties": "^7.10.1",
-        "@babel/preset-env": "^7.10.2",
-        "@babel/preset-react": "^7.10.1",
-        "babel-loader": "^8.1.0",
-        "css-loader": "^3.5.3",
-        "file-loader": "^6.0.0",
-        "style-loader": "^1.2.1",
-        "webpack": "^4.43.0",
-        "webpack-cli": "^3.3.11"
-    },
-    "prettier": {
-        "overrides": [
-            {
-                "files": [
-                    "*.js",
-                    "*.html",
-                    "*.json"
-                ],
-                "options": {
-                    "tabWidth": 4
-                }
-            }
-        ]
-    },
-    "repository": {
-        "type": "git",
-        "url": "git+https://gitlab.com/github.com/packit-service/dashboard.git"
-    },
-    "license": "MIT"
-}
