+++ packit_dashboard/home/routes.py
@@ -1,16 +1,15 @@
 from flask import Blueprint, render_template
-from packit_dashboard.config import API_URL
 
-home = Blueprint("home", __name__)
+home = Blueprint("home", __name__, template_folder="../../frontend/build")
 
 # Catch All URLs. Send to JS.
 
 
 @home.route("/")
 def main():
-    return render_template("index.html", api_url=API_URL)
+    return render_template("index.html")
 
 
 @home.route("/<path:path>")
 def catch_all(path):
-    return render_template("index.html", api_url=API_URL)
+    return render_template("index.html")
