--- templates/index.html
@@ -1,35 +1,0 @@
-<!DOCTYPE html>
-<html lang="en-US">
-    <head>
-        <meta charset="utf-8" />
-        <title>Packit Service</title>
-        <meta
-            id="packit-dashboard"
-            name="packit-service-dashboard"
-            content="Packit Service"
-        />
-        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-        <!-- TODO: Add Favicon -->
-        <link rel="icon" type="image/png" href="" />
-        <base href="/" />
-        <style>
-            html,
-            body,
-            #root {
-                height: 100%;
-            }
-        </style>
-    </head>
-
-    <body class="pf-m-redhat-font">
-        <noscript>Enabling JavaScript is required to run this app.</noscript>
-        <div id="root"></div>
-        <script>
-            const apiURL = "{{api_url}}";
-        </script>
-        <script
-            type="text/javascript"
-            src="{{ url_for('static', filename='dist/app.bundle.js') }}"
-        ></script>
-    </body>
-</html>
